test(libtonode): delete dead zcashd-era and duplicate test code
Bulk deletion pass (DRY refactor step R1):

- concrete.rs: the six commented-out zcashd-era tests
  (mempool_clearing_and_full_batch_syncs_correct_trees, witness_clearing,
  sandblast_filter_preserves_trees, standard/dust/shield_send_fees) and
  three redundant live tests: basic_scenario (every
  faucet_funded_recipient_default test covers it), note_selection_order
  (duplicate of the chain_generics fixture, own asserts commented out),
  and list_value_transfers_check_fees (strict subset of
  from_t_z_o_tz_to_zo_tzo_to_orchard; asserted nothing about value
  transfers).
- wallet.rs: deleted; its commented zcashd reorg test is dead and its one
  live test (verify_old_wallet_uses_server_height_in_send) moved verbatim
  into concrete.rs's fast module. One fewer test binary.
- shield_transparent.rs: deleted; ignored, assertion-free, subsumed by
  fixtures::send_shield_cycle and the coinbase-maturity tests.
- sync.rs: deleted the ignored assertion-free sync_test dev scratchpad.
- zingolib/src/testutils.rs: deleted the commented tx_inputs/tx_outputs/
  tx_actions/total_tx_value helpers and their orphaned
  TxNotesCount/TxActionsCount structs (no consumers workspace-wide).

Deliberately kept: mine_to_orchard / mine_to_sapling (zebrad can mine to
shielded pools) and the *_mobileclient scenarios (consumed externally by
zingo-mobile).
